feat: add context to date selector and change calender styles
