--- src/common/DatesSubHeader.tsx
@@ -1,479 +1,0 @@
-"use client";
-import { CompareDateItems, SelectDateItems } from "@/constants/SelectItems";
-import {
-  Button,
-  Checkbox,
-  Divider,
-  Input,
-  Select,
-  SelectItem,
-} from "@nextui-org/react";
-import useClickOutside from "hooks/useClickOutside";
-import React, { ChangeEvent, useState } from "react";
-import persian from "react-date-object/calendars/persian";
-import persian_fa from "react-date-object/locales/persian_fa";
-import { HiCalendarDays, HiOutlineInformationCircle } from "react-icons/hi2";
-import { Calendar, DateObject } from "react-multi-date-picker";
-
-const DatesSubHeader = () => {
-  const [showDatePicker, setShowDatePicker] = useState(false);
-  const datePickerRef = useClickOutside(() => setShowDatePicker(false));
-  const [isEnterUserDate, setIsEnterUserDate] = useState(true);
-  const [isCompare, setIsCompare] = useState(false);
-  const [startUserDate, setStartUserDate] = useState(
-    new DateObject({ calendar: persian }).subtract(1, "day").toString()
-  );
-  const [endUserDate, setEndUserDate] = useState(
-    new DateObject({ calendar: persian }).toString()
-  );
-  const [startCompareUserDate, setStartCompareUserDate] = useState(
-    new DateObject({ calendar: persian }).subtract(1, "day").toString()
-  );
-  const [endCompareUserDate, setEndCompareUserDate] = useState(
-    new DateObject({ calendar: persian }).toString()
-  );
-  const [desiredTimePeriod, setDesiredTimePeriod] = useState([
-    new DateObject({ calendar: persian }).toFirstOfWeek(),
-    new DateObject({ calendar: persian }).toLastOfWeek(),
-  ]);
-  const [compareTimePeriod, setCompareTimePeriod] = useState([
-    new DateObject({ calendar: persian }),
-    new DateObject({ calendar: persian }),
-  ]);
-  let startDesiredTimePeriod, endDesiredTimePeriod;
-  const SelectTimePeriodHandler = (e: ChangeEvent<HTMLSelectElement>) => {
-    const SelectId = e.target.value;
-    setIsEnterUserDate(false);
-    switch (SelectId) {
-      case "1":
-        setIsEnterUserDate(true);
-        return;
-      case "2":
-        startDesiredTimePeriod = new DateObject({ calendar: persian }).subtract(
-          2,
-          "day"
-        );
-        endDesiredTimePeriod = new DateObject({ calendar: persian }).subtract(
-          1,
-          "day"
-        );
-        setDesiredTimePeriod([startDesiredTimePeriod, endDesiredTimePeriod]);
-        return;
-      case "3":
-        let dayOfWeek = new DateObject({ calendar: persian }).weekDay.index;
-        let startOfThisWeek = new DateObject({ calendar: persian }).subtract(
-          dayOfWeek,
-          "days"
-        );
-        let endOfThisWeek = new DateObject({ calendar: persian }).subtract(
-          dayOfWeek,
-          "days"
-        );
-        startDesiredTimePeriod = startOfThisWeek.subtract(7, "days");
-        endDesiredTimePeriod = endOfThisWeek.subtract(1, "days");
-        setDesiredTimePeriod([startDesiredTimePeriod, endDesiredTimePeriod]);
-        return;
-      case "4":
-        let startLastMonthDate = new DateObject({ calendar: persian }).subtract(
-          1,
-          "month"
-        );
-        let endLastMonthDate = new DateObject({ calendar: persian }).subtract(
-          1,
-          "month"
-        );
-        startDesiredTimePeriod = startLastMonthDate.setDay(1);
-        endDesiredTimePeriod = endLastMonthDate.add(1, "month").setDay(1);
-        setDesiredTimePeriod([startDesiredTimePeriod, endDesiredTimePeriod]);
-        return;
-      case "5":
-        startDesiredTimePeriod = new DateObject({ calendar: persian }).subtract(
-          7,
-          "days"
-        );
-        endDesiredTimePeriod = new DateObject({ calendar: persian }).subtract(
-          1,
-          "day"
-        );
-        setDesiredTimePeriod([startDesiredTimePeriod, endDesiredTimePeriod]);
-        return;
-      case "6":
-        startDesiredTimePeriod = new DateObject({ calendar: persian }).subtract(
-          30,
-          "days"
-        );
-        endDesiredTimePeriod = new DateObject({ calendar: persian }).subtract(
-          1,
-          "day"
-        );
-        setDesiredTimePeriod([startDesiredTimePeriod, endDesiredTimePeriod]);
-        return;
-      case "7":
-        startDesiredTimePeriod = new DateObject({ calendar: persian }).subtract(
-          365,
-          "days"
-        );
-        endDesiredTimePeriod = new DateObject({ calendar: persian }).subtract(
-          1,
-          "day"
-        );
-        setDesiredTimePeriod([startDesiredTimePeriod, endDesiredTimePeriod]);
-        return;
-      default:
-        break;
-    }
-  };
-  const StartUserDateHandler = (e: ChangeEvent<HTMLInputElement>) => {
-    const inputValue = e.target.value;
-    setStartUserDate(inputValue);
-    startDesiredTimePeriod = new DateObject({
-      calendar: persian,
-      locale: persian_fa,
-      format: "YYYY/MM/DD",
-      date: startUserDate,
-    });
-  };
-  const EndUserDateHandler = (e: ChangeEvent<HTMLInputElement>) => {
-    const inputValue = e.target.value;
-    setEndUserDate(inputValue);
-    endDesiredTimePeriod = new DateObject({
-      calendar: persian,
-      locale: persian_fa,
-      format: "YYYY/MM/DD",
-      date: endUserDate,
-    });
-  };
-  const SplitDesiredTimePeriod = desiredTimePeriod.toString().split(",");
-  console.log(SplitDesiredTimePeriod);
-  let startCompareTimePeriod, endCompareTimePeriod;
-  const SelectCompareTimePeriodHandler = (e: ChangeEvent<HTMLSelectElement>) => {
-    const SelectId = e.target.value;
-     setIsEnterUserDate(false);
-    switch (SelectId) {
-      case "1":
-        setIsEnterUserDate(true);
-        return;
-      case "2":
-        let startCompareLastMonthDate = new DateObject(desiredTimePeriod[0]).subtract(1, "month");
-        let endCompareLastMonthDate = new DateObject(desiredTimePeriod[1]).subtract(1, "month");
-        startCompareTimePeriod = startCompareLastMonthDate.setDay(1);
-        endCompareTimePeriod = endCompareLastMonthDate.add(1, "month").setDay(1);
-         console.log(startCompareLastMonthDate.format("YYYY/MM/DD") , endCompareLastMonthDate.format("YYYY/MM/DD"));
-         setCompareTimePeriod([startCompareTimePeriod , endCompareTimePeriod]);
-          return;
-      
-       case "3":
-         let startCompareLastYearDate = new DateObject(desiredTimePeriod[0]).subtract(1, "year");
-         let endCompareLastYearDate = new DateObject(desiredTimePeriod[1]).subtract(1, "year");
-         startCompareTimePeriod = startCompareLastYearDate.setDay(1);
-         endCompareTimePeriod = endCompareLastYearDate.add(1, "year").setDay(1);
-         console.log(startCompareLastYearDate.format("YYYY/MM/DD") , endCompareLastYearDate.format("YYYY/MM/DD"));
-         setCompareTimePeriod([startCompareTimePeriod , endCompareTimePeriod]);
-        return;
-      default:
-        break;
-    }
-  };
-  const StartCompareUserDateHandler = (e: ChangeEvent<HTMLInputElement>) => {
-    const inputValue = e.target.value;
-    setStartCompareUserDate(inputValue);
-    startCompareTimePeriod = new DateObject({
-      calendar: persian,
-      locale: persian_fa,
-      format: "YYYY/MM/DD",
-      date: startCompareUserDate,
-    });
-  };
-  const EndCompareUserDateHandler = (e: ChangeEvent<HTMLInputElement>) => {
-    const inputValue = e.target.value;
-    setEndCompareUserDate(inputValue);
-    endCompareTimePeriod = new DateObject({
-      calendar: persian,
-      locale: persian_fa,
-      format: "YYYY/MM/DD",
-      date: endCompareUserDate,
-    });
-  };
-  const SplitCompareTimePeriod = compareTimePeriod.toString().split(",");
-  console.log(SplitCompareTimePeriod)
-  return (
-    <section className="relative w-full flex-between bg-white h-20 mb-3 rounded-lg shadow-fullLight">
-      <div className="flex-center h-full">
-        {/* Trigger Btn */}
-        <Button
-          color="primary"
-          variant="light"
-          onPress={() => setShowDatePicker((prev) => !prev)}
-          className="hover:bg-primary-50 min-w-52 h-full px-1"
-        >
-          <div className="flex flex-col text-xs md:text-sm gap-y-2">
-            <div className="flex-center gap-x-1">
-              <HiCalendarDays className="size-4" />
-              <span>
-                {isEnterUserDate ? startUserDate : SplitDesiredTimePeriod[0]}
-              </span>
-              <span>تا</span>
-              <span>
-                {isEnterUserDate ? endUserDate : SplitDesiredTimePeriod[1]}
-              </span>
-            </div>
-             <div className={`${isCompare ? "opacity-100 visible" : "opacity-0 invisible"} flex-center gap-x-1 transition-all ease-in-out duration-30`}>
-                  <span>مقایسه با</span>
-                  <span>{isEnterUserDate ? startCompareUserDate : SplitCompareTimePeriod[0]}</span>
-                  <span>تا</span>
-                  <span>{isEnterUserDate ? endCompareUserDate : SplitCompareTimePeriod[1]}</span> 
-                  </div>
-            
-            </div>
-        </Button>
-        <Divider orientation="vertical" className="h-full w-px bg-gray-100" />
-        <div
-          ref={datePickerRef}
-          className={`${
-            !showDatePicker ? "opacity-100 visible" : "opacity-0 invisible"
-          } absolute z-50 w-full flex flex-col max-w-[65rem] right-0 top-20 rounded-lg bg-white overflow-hidden shadow-fullDark transition-all ease-in-out duration-300`}
-        >
-          {/* Main Container */}
-          <div className="flex h-full min-h-[29rem]">
-            {/* Left Side */}
-            <div className="w-2/5 border-l border-gray-100 text-zinc-700">
-               {/* Left Side Top */}
-              <div className="h-1/2 border-r-4 border-r-primary overflow-hidden border-b border-b-gray-100 px-3 py-4">
-                <p className="my-3">انتخاب بازه زمانی</p>
-                <Select
-                  size="lg"
-                  items={SelectDateItems}
-                  aria-label="انتخاب بازه زمانی"
-                  aria-labelledby="انتخاب بازه زمانی"
-                  defaultSelectedKeys="1"
-                  onChange={SelectTimePeriodHandler}
-                  variant="bordered"
-                  classNames={{ base: "my-5" }}
-                >
-                  {(SelectDateItem) => (
-                    <SelectItem
-                      key={SelectDateItem.id}
-                      value={SelectDateItem.title}
-                    >
-                      {SelectDateItem.title}
-                    </SelectItem>
-                  )}
-                </Select>
-                <div className="flex-between gap-x-4 my-6">
-                  {isEnterUserDate ? (
-                    <>
-                      <Input
-                        size="lg"
-                        variant="bordered"
-                        placeholder="تاریخ متعبر: ۱۴۰۱/۰۱/۰۱"
-                        value={startUserDate.toString()}
-                        onChange={StartUserDateHandler}
-                        classNames={{
-                          inputWrapper:
-                            "data-[hover=true]:border-primary group-data-[focus=true]:border-primary",
-                          input: "text-center text-base placeholder:text-xs",
-                        }}
-                      />
-                      <span>تا</span>
-                      <Input
-                        size="lg"
-                        variant="bordered"
-                        placeholder="تاریخ متعبر: ۱۴۰۱/۰۲/۰۲"
-                        value={endUserDate.toString()}
-                        onChange={EndUserDateHandler}
-                        classNames={{
-                          inputWrapper:
-                            "data-[hover=true]:border-primary group-data-[focus=true]:border-primary",
-                          input: "text-center text-base placeholder:text-xs",
-                        }}
-                      />
-                    </>
-                  ) : (
-                    <>
-                      <Input
-                        size="lg"
-                        variant="bordered"
-                        value={SplitDesiredTimePeriod[0]}
-                        classNames={{
-                          inputWrapper:
-                            "data-[hover=true]:border-primary group-data-[focus=true]:border-primary",
-                          input: "text-center text-base",
-                        }}
-                      />
-                      <span>تا</span>
-                      <Input
-                        size="lg"
-                        variant="bordered"
-                        value={SplitDesiredTimePeriod[1]}
-                        classNames={{
-                          inputWrapper:
-                            "data-[hover=true]:border-primary group-data-[focus=true]:border-primary",
-                          input: "text-center text-base",
-                        }}
-                      />
-                    </>
-                  )}
-                </div>
-              </div>
-              {/* Left Side Bottom */}
-              <div
-                className={`${
-                  isCompare && "border-r-4 border-r-amber-500"
-                } h-1/2  text-zinc-700 overflow-hidden border-b border-b-gray-100 px-3 py-4`}
-              >
-                <Checkbox
-                  isSelected={isCompare}
-                  onChange={() => setIsCompare((prev) => !prev)}
-                  color="primary"
-                  radius="lg"
-                  className="my-3"
-                >
-                  <span className="text-zinc-700">مقایسه با ...</span>
-                </Checkbox>
-                <Select
-                  isDisabled={!isCompare && true}
-                  size="lg"
-                  items={CompareDateItems}
-                  aria-label=" مقایسه با "
-                  aria-labelledby="مقایسه با"
-                  defaultSelectedKeys="1"
-                  onChange={SelectCompareTimePeriodHandler}
-                  variant="bordered"
-                  classNames={{ base: "my-5" }}
-                >
-                  {(CompareDateItem) => (
-                    <SelectItem
-                      key={CompareDateItem.id}
-                      value={CompareDateItem.title}
-                    >
-                      {CompareDateItem.title}
-                    </SelectItem>
-                  )}
-                </Select>
-                <div className="flex-between gap-x-4 my-6">
-                  {isEnterUserDate ? (
-                    <>
-                      <Input
-                        isDisabled={!isCompare && true}
-                        size="lg"
-                        variant="bordered"
-                        placeholder="تاریخ متعبر: ۱۴۰۱/۰۱/۰۱"
-                        value={startCompareUserDate.toString()}
-                        onChange={StartCompareUserDateHandler}
-                        classNames={{
-                          inputWrapper:
-                            "data-[hover=true]:border-primary group-data-[focus=true]:border-primary",
-                          input: "text-center text-base placeholder:text-xs",
-                        }}
-                      />
-                      <span
-                        className={`${
-                          !isCompare ? "opacity-50" : "opacity-100"
-                        }`}
-                      >
-                        تا
-                      </span>
-                      <Input
-                        isDisabled={!isCompare && true}
-                        size="lg"
-                        variant="bordered"
-                        placeholder="تاریخ متعبر: ۱۴۰۱/۰۲/۰۲"
-                        value={endCompareUserDate.toString()}
-                        onChange={EndCompareUserDateHandler}
-                        classNames={{
-                          inputWrapper:
-                            "data-[hover=true]:border-primary group-data-[focus=true]:border-primary",
-                          input: "text-center text-base placeholder:text-xs",
-                        }}
-                      />
-                    </>
-                  ) : (
-                    <>
-                      <Input
-                        isDisabled={!isCompare && true}
-                        size="lg"
-                        variant="bordered"
-                        value={SplitCompareTimePeriod[0]}
-                        classNames={{
-                          inputWrapper:
-                            "data-[hover=true]:border-primary group-data-[focus=true]:border-primary",
-                          input: "text-center text-base",
-                        }}
-                      />
-                      <span
-                        className={`${
-                          !isCompare ? "opacity-50" : "opacity-100"
-                        }`}
-                      >
-                        تا
-                      </span>
-                      <Input
-                        isDisabled={!isCompare && true}
-                        size="lg"
-                        variant="bordered"
-                        value={SplitCompareTimePeriod[1]}
-                        classNames={{
-                          inputWrapper:
-                            "data-[hover=true]:border-primary group-data-[focus=true]:border-primary",
-                          input: "text-center text-base",
-                        }}
-                      />
-                    </>
-                  )}
-                </div>
-              </div>
-            </div>
-            {/* Right Side */}
-            <div className="w-3/5 min-h-[23rem] px-3 py-4">
-              <div className="flex-center">
-                <Calendar
-                  range
-                  rangeHover
-                  value={desiredTimePeriod}
-                  calendar={persian}
-                  locale={persian_fa}
-                  onChange={setDesiredTimePeriod}
-                />
-                <Calendar
-                  range
-                  disabled={!isCompare && true}
-                  rangeHover
-                  value={compareTimePeriod}
-                  onChange={setCompareTimePeriod}
-                  calendar={persian}
-                  locale={persian_fa}
-                />
-              </div>
-            </div>
-          </div>
-          <Divider />
-          {/* Buttons */}
-          <div className="w-full flex justify-end items-center p-6">
-            <div className="flex-center gap-x-4">
-              <Button
-                color="primary"
-                variant="bordered"
-                onPress={() => setShowDatePicker(false)}
-              >
-                انصراف
-              </Button>
-              <Button
-                color="primary"
-                variant="solid"
-                onPress={() => setShowDatePicker(true)}
-              >
-                اعمال
-              </Button>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="flex-center font-normal h-full gap-x-1 pl-1.5">
-        <HiOutlineInformationCircle className="size-4" />
-        <span>راهنما</span>
-      </div>
-    </section>
-  );
-};
-
-export default DatesSubHeader;
